CLHEP: Sync with CLHEP 2.4.6.3

+++ source/externals/clhep/src/CLHEP/Matrix/SymMatrix.cc
@@ -676,7 +676,7 @@ std::ostream& operator<<(std::ostream &os, const HepSymMatrix &q)
 {
   os << std::endl;
 /* Fixed format needs 3 extra characters for field, while scientific needs 7 */
-  int width;
+  long width;
   if(os.flags() & std::ios::fixed)
     width = os.precision()+3;
   else
