+++ source/externals/clhep/src/CLHEP/Matrix/Vector.cc
@@ -514,7 +514,7 @@ std::ostream& operator<<(std::ostream &os, const HepVector &q)
 {
   os << std::endl;
 /* Fixed format needs 3 extra characters for field, while scientific needs 7 */
-  int width;
+  long width;
   if(os.flags() & std::ios::fixed)
     width = os.precision()+3;
   else
